pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=2.0.0,<3.0.0"]

[dependency-groups]
dev = [
  "pylint-pydantic (>=0.4.1,<0.5.0)"
]

[project]
authors = [
  {email = "zhulikovnikita884@gmail.com", name = "nikitazulikov"}
]
dependencies = [
  "aiohttp (>=3.13.2,<4.0.0)",
  "amocrm-api (>=2.6.1,<3.0.0)",
  "black (>=25.9.0,<26.0.0)",
  "dotenv (>=0.9.9,<0.10.0)",
  "fastapi (>=0.120.2,<0.121.0)",
  "flake8 (>=7.3.0,<8.0.0)",
  "flake8-pyproject (>=1.2.3,<2.0.0)",
  "google-auth (>=2.42.0,<3.0.0)",
  "gspread (>=6.2.1,<7.0.0)",
  "isort (>=7.0.0,<8.0.0)",
  "mypy (>=1.18.2,<2.0.0)",
  "pydantic-settings (>=2.11.0,<3.0.0)",
  "pylint (>=4.0.2,<5.0.0)",
  "pytest (>=8.4.2,<9.0.0)",
  "python-multipart (>=0.0.20,<0.0.21)",
  "redis (>=7.0.1,<8.0.0)",
  "tenacity (>=9.1.2,<10.0.0)",
  "toml-sort (>=0.24.3,<0.25.0)",
  "uvicorn (>=0.38.0,<0.39.0)"
]
description = ""
name = "amocrm-gsheets-integration"
readme = "README.md"
requires-python = ">=3.13"
version = "0.1.0"

[tool.flake8]
extend-ignore = ["E203", "W503"]
max-line-length = 128
